change option max value

+++ runtime/thread-pool.cpp
@@ -26,4 +26,4 @@ void ThreadPool::init() noexcept {
   }
 }
 
-uint64_t thread_pool_size = 0;
+int thread_pool_size = 0;
